requirements.txt
streamlit==1.37.1
streamlit-option-menu==0.3.12
requests
pandas
openpyxl
numpy
gTTS==2.5.1
moviepy==1.0.3
decorator==4.4.2
imageio==2.35.1
imageio-ffmpeg==0.5.1
proglog==0.1.10
tqdm==4.66.5
reportlab
pillow
